feat(polars): remove wrapper, fix docstring, re-export
