+++ Cargo.toml
@@ -34,7 +34,7 @@ vtab-full = ["vtab-excel", "vtab-arrow"]
 extensions-full = ["httpfs", "json", "parquet", "vtab-full"]
 buildtime_bindgen = ["libduckdb-sys/buildtime_bindgen"]
 modern-full = ["chrono", "serde_json", "url", "r2d2", "uuid", "polars"]
-polars = ["dep:arrow2", "dep:polars", "dep:polars-core"]
+polars = ["dep:arrow2", "dep:polars"]
 
 [dependencies]
 # time = { version = "0.3.2", features = ["formatting", "parsing"], optional = true }
@@ -60,7 +60,6 @@ num = { version = "0.4", optional = true, default-features = false, features = [
 duckdb-loadable-macros = { version = "0.1.0", path="./duckdb-loadable-macros", optional = true }
 polars = { version = "0.30.0", features = ["dtype-full"], optional = true}
 arrow2 = { version = "0.17.2", features = ["arrow"], optional = true }
-polars-core = { version = "0.30.0", optional = true }
 
 [dev-dependencies]
 doc-comment = "0.3"
@@ -71,6 +70,7 @@ uuid = { version = "1.0", features = ["v4"] }
 unicase = "2.6.0"
 rand = "0.8.3"
 tempdir = "0.3.7"
+polars-core = "0.30.0"
 # criterion = "0.3"
 
 # [[bench]]
